# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwissGold.cz precious metals e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List active products
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: subcat
          in: query
          required: false
          schema: { type: string }
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: featured
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    post:
      operationId: createOrder
      tags: [shop]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login, returns a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: List all products (incl. inactive)
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: List orders
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Get order with items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: adminUpdateOrder
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/buybacks:
    get:
      operationId: adminListBuybacks
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: List buyback requests
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Buybacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Buyback"
  /admin/buybacks/{id}:
    get:
      operationId: adminGetBuyback
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Get buyback request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Buyback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Buyback"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: adminUpdateBuyback
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Update buyback status and note
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuybackStatusUpdate"
      responses:
        "200":
          description: Updated buyback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Buyback"
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Get settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /admin/content:
    get:
      operationId: adminListContent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: List all content blocks
      responses:
        "200":
          description: Content blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentBlock"
  /admin/content/{key}:
    put:
      operationId: adminUpdateContent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Update a content block
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentBlockUpdate"
      responses:
        "200":
          description: Updated content block
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentBlock"
  /admin/overrides:
    get:
      operationId: adminListOverrides
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: List price overrides
      responses:
        "200":
          description: Overrides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceOverride"
  /admin/overrides/{id}:
    put:
      operationId: adminUpdateOverride
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Create or update a price override
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceOverrideUpdate"
      responses:
        "200":
          description: Updated override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceOverride"
  /admin/customers:
    get:
      operationId: adminListCustomers
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: List customers with order/buyback counts
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCustomerSummary"
  /admin/customers/{id}:
    get:
      operationId: adminGetCustomer
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Get a customer with their orders and buybacks
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: sg_customer_token
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - weightGrams
        - fineness
        - category
        - subcat
        - featured
        - active
        - sortOrder
      properties:
        id: { type: string }
        name: { type: string }
        manufacturer: { type: ["string", "null"] }
        weightGrams: { type: number }
        fineness: { type: string }
        category: { type: string }
        subcat: { type: string }
        year: { type: ["integer", "null"] }
        featured: { type: boolean }
        active: { type: boolean }
        image: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        sortOrder: { type: integer }
    ProductUpdate:
      type: object
      required:
        - name
        - category
        - subcat
        - weightGrams
        - fineness
        - featured
        - active
        - sortOrder
      properties:
        name: { type: string, minLength: 1 }
        manufacturer: { type: ["string", "null"] }
        weightGrams: { type: number }
        fineness: { type: string }
        category: { type: string }
        subcat: { type: string }
        year: { type: ["integer", "null"] }
        featured: { type: boolean }
        active: { type: boolean }
        image: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        sortOrder: { type: integer }
...
    ContentBlock:
      type: object
      required:
        - key
        - value
      properties:
        key: { type: string }
        value: { type: string }
    ContentBlockUpdate:
      type: object
      required:
        - value
      properties:
        value: { type: string }
    OrderItemInput:
      type: object
      required:
        - productId
        - productName
        - quantity
        - unitPriceCzk
        - unitPriceEur
      properties:
        productId: { type: string }
        productName: { type: string }
        weightGrams: { type: ["number", "null"] }
        quantity: { type: integer }
        unitPriceCzk: { type: number }
        unitPriceEur: { type: number }
    OrderInput:
      type: object
      required:
        - customerName
        - customerEmail
        - paymentMethod
        - deliveryMethod
        - currency
        - totalCzk
        - totalEur
        - items
      properties:
        customerName: { type: string, minLength: 1 }
        customerEmail: { type: string, minLength: 1 }
        customerPhone: { type: ["string", "null"] }
        customerAddress: { type: ["string", "null"] }
        customerCity: { type: ["string", "null"] }
        customerZip: { type: ["string", "null"] }
        paymentMethod: { type: string }
        deliveryMethod: { type: string }
        currency: { type: string }
        totalCzk: { type: number }
        totalEur: { type: number }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItem:
      type: object
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPriceCzk
        - unitPriceEur
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        weightGrams: { type: ["number", "null"] }
        quantity: { type: integer }
        unitPriceCzk: { type: number }
        unitPriceEur: { type: number }
    Order:
      type: object
      required:
        - id
        - orderNumber
        - status
        - customerName
        - customerEmail
        - paymentMethod
        - deliveryMethod
        - totalCzk
        - totalEur
        - currency
        - createdAt
        - items
      properties:
        id: { type: string }
        orderNumber: { type: string }
        status: { type: string }
        customerId: { type: ["string", "null"] }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: ["string", "null"] }
        customerAddress: { type: ["string", "null"] }
        customerCity: { type: ["string", "null"] }
        customerZip: { type: ["string", "null"] }
        paymentMethod: { type: string }
        deliveryMethod: { type: string }
        totalCzk: { type: number }
        totalEur: { type: number }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderStatusUpdate:
      type: object
      required:
        - status
      properties:
        status: { type: string }
...
    Buyback:
      type: object
      required:
        - id
        - requestNumber
        - status
        - customerName
        - customerEmail
        - createdAt
      properties:
        id: { type: string }
        requestNumber: { type: string }
        status: { type: string }
        customerId: { type: ["string", "null"] }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: ["string", "null"] }
        itemDescription: { type: ["string", "null"] }
        estimatedCzk: { type: ["number", "null"] }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    BuybackStatusUpdate:
      type: object
      required:
        - status
      properties:
        status: { type: string }
        adminNote: { type: ["string", "null"] }
    Settings:
      type: object
      required:
        - id
        - eurToCzk
        - globalMarginCzk
        - marginInvestZlato
        - marginInvestStribro
        - marginPlatinaPalladium
        - marginMinceCnb
        - buybackSpreadPct
        - deferredDiscountPct
        - bulkTier1Qty
        - bulkTier1DiscountPct
        - bulkTier2Qty
        - bulkTier2DiscountPct
        - bulkTier3Qty
        - bulkTier3DiscountPct
      properties:
        id: { type: integer }
        eurToCzk: { type: number }
        globalMarginCzk: { type: number }
        marginInvestZlato: { type: number }
        marginInvestStribro: { type: number }
        marginPlatinaPalladium: { type: number }
        marginMinceCnb: { type: number }
        buybackSpreadPct: { type: number }
        deferredDiscountPct: { type: number }
        bulkTier1Qty: { type: integer }
        bulkTier1DiscountPct: { type: number }
        bulkTier2Qty: { type: integer }
        bulkTier2DiscountPct: { type: number }
        bulkTier3Qty: { type: integer }
        bulkTier3DiscountPct: { type: number }
    SettingsUpdate:
      type: object
      properties:
        eurToCzk: { type: number }
        globalMarginCzk: { type: number }
        marginInvestZlato: { type: number }
        marginInvestStribro: { type: number }
        marginPlatinaPalladium: { type: number }
        marginMinceCnb: { type: number }
        buybackSpreadPct: { type: number }
        deferredDiscountPct: { type: number }
        bulkTier1Qty: { type: integer }
        bulkTier1DiscountPct: { type: number }
        bulkTier2Qty: { type: integer }
        bulkTier2DiscountPct: { type: number }
        bulkTier3Qty: { type: integer }
        bulkTier3DiscountPct: { type: number }
    PriceOverride:
      type: object
      required:
        - productId
        - active
      properties:
        productId: { type: string }
        marginCzk: { type: ["number", "null"] }
        marginPct: { type: ["number", "null"] }
        active: { type: boolean }
    PriceOverrideUpdate:
      type: object
      required:
        - active
      properties:
        marginCzk: { type: ["number", "null"] }
        marginPct: { type: ["number", "null"] }
        active: { type: boolean }
    AdminCredentials:
      type: object
      required:
        - email
        - password
      properties:
        email: { type: string }
        password: { type: string }
    AuthToken:
      type: object
      required:
        - token
      properties:
        token: { type: string }
    AdminStats:
      type: object
      required:
        - productCount
        - newOrders
        - newBuybacks
        - spotGoldCzkPerGram
        - eurCzk
      properties:
        productCount: { type: integer }
        newOrders: { type: integer }
        newBuybacks: { type: integer }
        spotGoldCzkPerGram: { type: number }
        eurCzk: { type: number }
    Customer:
      type: object
      required:
        - id
        - email
        - createdAt
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        zip: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    AdminCustomerSummary:
      type: object
      required:
        - id
        - email
        - createdAt
        - orderCount
        - totalSpentCzk
        - buybackCount
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        orderCount: { type: integer }
        totalSpentCzk: { type: number }
        buybackCount: { type: integer }
    AdminCustomerDetail:
      type: object
      required:
        - customer
        - orders
        - buybacks
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        buybacks:
          type: array
          items:
            $ref: "#/components/schemas/Buyback"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error: { type: string }