# lib/api-zod/src/generated/types/adminCustomerDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwissGold.cz precious metals e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Buyback } from './buyback';
import type { Customer } from './customer';
import type { Order } from './order';

export interface AdminCustomerDetail {
  customer: Customer;
  orders: Order[];
  buybacks: Buyback[];
}