# lib/api-zod/src/generated/types/adminCustomerSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwissGold.cz precious metals e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCustomerSummary {
  id: string;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: Date;
  orderCount: number;
  totalSpentCzk: number;
  buybackCount: number;
}