# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwissGold.cz precious metals e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: string;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  zip?: string | null;
  createdAt: Date;
}