# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwissGold.cz precious metals e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: string;
  orderNumber: string;
  status: string;
  /** @nullable */
  customerId?: string | null;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  customerCity?: string | null;
  /** @nullable */
  customerZip?: string | null;
  paymentMethod: string;
  deliveryMethod: string;
  totalCzk: number;
  totalEur: number;
  currency: string;
  createdAt: Date;
  items: OrderItem[];
}